+++ sop-common/sop-gateway-common/src/main/java/com/gitee/sop/gatewaycommon/result/ResultExecutorForZuul.java
@@ -1,9 +1,0 @@
-package com.gitee.sop.gatewaycommon.result;
-
-import com.netflix.zuul.context.RequestContext;
-
-/**
- * @author tanghc
- */
-public interface ResultExecutorForZuul extends ResultExecutor<RequestContext, String> {
-}