+++ sop-common/sop-service-common/src/main/java/com/gitee/sop/servercommon/bean/ServiceApiInfo.java
@@ -34,10 +34,6 @@ public class ServiceApiInfo {
         private int permission;
         /** 是否需要token */
         private int needToken;
-        /** 是否是兼容模式，即使用了@ApiAbility注解 */
-        private int compatibleMode;
-        /** 是否是原始Mapping */
-        private boolean originalMapping;
 
         public ApiMeta() {
         }