+++ sop-common/sop-service-common/src/main/java/com/gitee/sop/servercommon/bean/ServiceContext.java
@@ -12,10 +12,6 @@ public class ServiceContext extends ConcurrentHashMap<String, Object> {
 
     public static final String REQUEST_KEY = "request";
     public static final String RESPONSE_KEY = "response";
-    public static final String OPEN_CONTEXT_KEY = "openContext";
-    public static final String SOP_MVC = "sop.mvc";
-
-    public static final String RESTFUL_KEY = "sop-restful";
 
     protected static Class<? extends ServiceContext> contextClass = ServiceContext.class;
 
@@ -33,14 +29,6 @@ public class ServiceContext extends ConcurrentHashMap<String, Object> {
 
     public ServiceContext() {
         super();
-    }
-
-    public void setOpenContext(OpenContext<?> openContext) {
-        this.set(OPEN_CONTEXT_KEY, openContext);
-    }
-
-    public OpenContext getOpenContext() {
-        return (OpenContext) this.get(OPEN_CONTEXT_KEY);
     }
 
     public Locale getLocale() {