+++ sop-test/src/test/java/com/gitee/sop/test/AlipayClientPostTest.java
@@ -36,12 +36,10 @@ public class AlipayClientPostTest extends TestBase {
     @Test
     public void testGet() throws Exception {
 
-        // 接口实现见：com.gitee.sop.storyweb.controller.AlipayController.getStory
-
         // 公共请求参数
         Map<String, String> params = new HashMap<String, String>();
         params.put("app_id", appId);
-        params.put("method", "alipay.story.get");
+        params.put("method", "story.get");
         params.put("format", "json");
         params.put("charset", "utf-8");
         params.put("sign_type", "RSA2");